Cleaning up rig.js
